Refactor: Split dataLib.js into multiple focused library files
