+++ ITEM RARITY/passivesLib.js
@@ -1,0 +1,12 @@
+export const passives = {
+    FROST_TOUCH: {
+        name: "§bFrost Touch",
+        type: "sword",
+        rarity: "Rare",
+        min: 1,
+        max: 2,
+        description: "slow enemy for §x{x} seconds on hit",
+        cooldown: 10,
+        scoreboard: "frosttouch"
+    }
+};
