+++ ITEM RARITY/skillsLib.js
@@ -1,0 +1,49 @@
+export const skills = {
+    // SMASH LEAP - Mace
+    SMASH_LEAP_RARE: { name: "§9Smash Leap", type: "mace", rarity: "Rare", min: 2, max: 2, description: "§7Smashes the ground and stuns \n§7enemies in §x{x} §7blocks radius, \n§7then leaps 8 blocks if has wind charge", cooldown: 120, scoreboard: "smashleap" },
+    SMASH_LEAP_EPIC: { name: "§5Smash Leap", type: "mace", rarity: "Epic", min: 2, max: 3, description: "§7Smashes the ground and stuns \n§7enemies in §x{x} §7blocks radius, \n§7then leaps 8 blocks if has wind charge", cooldown: 120, scoreboard: "smashleap" },
+    SMASH_LEAP_LEGENDARY: { name: "§6Smash Leap", type: "mace", rarity: "Legendary", min: 3, max: 5, description: "§7Smashes the ground and stuns \n§7enemies in §x{x} §7blocks radius, \n§7then leaps 8 blocks if has wind charge", cooldown: 120, scoreboard: "smashleap" },
+    SMASH_LEAP_MYTHIC: { name: "§cSmash Leap", type: "mace", rarity: "Mythic", min: 5, max: 7, description: "§7Smashes the ground and stuns \n§7enemies in §x{x} §7blocks radius, \n§7then leaps 8 blocks if has wind charge", cooldown: 120, scoreboard: "smashleap" },
+
+    // SPIN STRIKE - Sword
+    SPIN_STRIKE_RARE: { name: "§9Spin Strike", type: "sword", rarity: "Rare", min: 5, max: 6, description: "§7Performs a 360° spin attack, \n§7dealing §x{x} §7damage to all \n§7enemies within 3 blocks", cooldown: 60, scoreboard: "spinstrike" },
+    SPIN_STRIKE_EPIC: { name: "§5Spin Strike", type: "sword", rarity: "Epic", min: 7, max: 9, description: "§7Performs a 360° spin attack, \n§7dealing §x{x} §7damage to all \n§7enemies within 3 blocks", cooldown: 60, scoreboard: "spinstrike" },
+    SPIN_STRIKE_LEGENDARY: { name: "§6Spin Strike", type: "sword", rarity: "Legendary", min: 10, max: 15, description: "§7Performs a 360° spin attack, \n§7dealing §x{x} §7damage to all \n§7enemies within 3 blocks", cooldown: 60, scoreboard: "spinstrike" },
+    SPIN_STRIKE_MYTHIC: { name: "§cSpin Strike", type: "sword", rarity: "Mythic", min: 15, max: 20, description: "§7Performs a 360° spin attack, \n§7dealing §x{x} §7damage to all \n§7enemies within 3 blocks", cooldown: 60, scoreboard: "spinstrike" },
+
+    // EXPLOSIVE MINING - Pickaxe
+    EXPLOSIVE_MINING_RARE: { name: "§9Explosive Mining", type: "pickaxe", rarity: "Rare", min: 2, max: 3, description: "§7Creates an explosion with \n§7size of §x{x} §7to break \n§7blocks and damage enemies", cooldown: 1000, scoreboard: "explosivemining" },
+    EXPLOSIVE_MINING_EPIC: { name: "§5Explosive Mining", type: "pickaxe", rarity: "Epic", min: 4, max: 5, description: "§7Creates an explosion with \n§7size of §x{x} §7to break \n§7blocks and damage enemies", cooldown: 1000, scoreboard: "explosivemining" },
+    EXPLOSIVE_MINING_LEGENDARY: { name: "§6Explosive Mining", type: "pickaxe", rarity: "Legendary", min: 6, max: 7, description: "§7Creates an explosion with \n§7size of §x{x} §7to break \n§7blocks and damage enemies", cooldown: 1000, scoreboard: "explosivemining" },
+    EXPLOSIVE_MINING_MYTHIC: { name: "§cExplosive Mining", type: "pickaxe", rarity: "Mythic", min: 8, max: 10, description: "§7Creates an explosion with \n§7size of §x{x} §7to break \n§7blocks and damage enemies", cooldown: 1000, scoreboard: "explosivemining" },
+
+    // RAY MINER - Pickaxe
+    RAY_MINER_RARE: { name: "§9Ray Miner", type: "pickaxe", rarity: "Rare", min: 8, max: 12, description: "§7Breaks §x{x} §7blocks in a \n§7straight line through solid \n§7materials", cooldown: 180, scoreboard: "rayminer" },
+    RAY_MINER_EPIC: { name: "§5Ray Miner", type: "pickaxe", rarity: "Epic", min: 12, max: 16, description: "§7Breaks §x{x} §7blocks in a \n§7straight line through solid \n§7materials", cooldown: 180, scoreboard: "rayminer" },
+    RAY_MINER_LEGENDARY: { name: "§6Ray Miner", type: "pickaxe", rarity: "Legendary", min: 16, max: 20, description: "§7Breaks §x{x} §7blocks in a \n§7straight line through solid \n§7materials", cooldown: 180, scoreboard: "rayminer" },
+    RAY_MINER_MYTHIC: { name: "§cRay Miner", type: "pickaxe", rarity: "Mythic", min: 20, max: 25, description: "§7Breaks §x{x} §7blocks in a \n§7straight line through solid \n§7materials", cooldown: 180, scoreboard: "rayminer" },
+
+    // EXCAVATOR - Shovel
+    EXCAVATOR_RARE: { name: "§9Excavator", type: "shovel", rarity: "Rare", min: 2, max: 4, description: "§7Breaks sand-type blocks in a \n§73x3x§x{x} §7area, consuming \n§7durability", cooldown: 70, scoreboard: "excavator" },
+    EXCAVATOR_EPIC: { name: "§5Excavator", type: "shovel", rarity: "Epic", min: 4, max: 6, description: "§7Breaks sand-type blocks in a \n§73x3x§x{x} §7area, consuming \n§7durability", cooldown: 70, scoreboard: "excavator" },
+    EXCAVATOR_LEGENDARY: { name: "§6Excavator", type: "shovel", rarity: "Legendary", min: 6, max: 8, description: "§7Breaks sand-type blocks in a \n§73x3x§x{x} §7area, consuming \n§7durability", cooldown: 70, scoreboard: "excavator" },
+    EXCAVATOR_MYTHIC: { name: "§cExcavator", type: "shovel", rarity: "Mythic", min: 7, max: 12, description: "§7Breaks sand-type blocks in a \n§73x3x§x{x} §7area, consuming \n§7durability", cooldown: 70, scoreboard: "excavator" },
+
+    // FLAME ARC - Sword
+    FLAME_ARC_RARE: { name: "§9Flame Arc", type: "sword", rarity: "Rare", min: 15, max: 20, description: "§7Unleashes an arc of fire that \n§7ignites enemies and blocks, \n§7dealing §x{x} §7fire damage", cooldown: 400, scoreboard: "flamearc" },
+    FLAME_ARC_EPIC: { name: "§5Flame Arc", type: "sword", rarity: "Epic", min: 20, max: 28, description: "§7Unleashes an arc of fire that \n§7ignites enemies and blocks, \n§7dealing §x{x} §7fire damage", cooldown: 400, scoreboard: "flamearc" },
+    FLAME_ARC_LEGENDARY: { name: "§6Flame Arc", type: "sword", rarity: "Legendary", min: 28, max: 38, description: "§7Unleashes an arc of fire that \n§7ignites enemies and blocks, \n§7dealing §x{x} §7fire damage", cooldown: 400, scoreboard: "flamearc" },
+    FLAME_ARC_MYTHIC: { name: "§cFlame Arc", type: "sword", rarity: "Mythic", min: 38, max: 50, description: "§7Unleashes an arc of fire that \n§7ignites enemies and blocks, \n§7dealing §x{x} §7fire damage", cooldown: 400, scoreboard: "flamearc" },
+
+    // SHADOW DASH - Sword
+    SHADOW_DASH_RARE: { name: "§9Shadow Dash", type: "sword", rarity: "Rare", min: 3, max: 5, description: "§7Dash forward §x{x} §7blocks \n§7through enemies, reduced \n§7distance when airborne", cooldown: 50, scoreboard: "shadowdash" },
+    SHADOW_DASH_EPIC: { name: "§5Shadow Dash", type: "sword", rarity: "Epic", min: 6, max: 8, description: "§7Dash forward §x{x} §7blocks \n§7through enemies, reduced \n§7distance when airborne", cooldown: 50, scoreboard: "shadowdash" },
+    SHADOW_DASH_LEGENDARY: { name: "§6Shadow Dash", type: "sword", rarity: "Legendary", min: 9, max: 12, description: "§7Dash forward §x{x} §7blocks \n§7through enemies, reduced \n§7distance when airborne", cooldown: 50, scoreboard: "shadowdash" },
+    SHADOW_DASH_MYTHIC: { name: "§cShadow Dash", type: "sword", rarity: "Mythic", min: 13, max: 15, description: "§7Dash forward §x{x} §7blocks \n§7through enemies, reduced \n§7distance when airborne", cooldown: 50, scoreboard: "shadowdash" },
+
+    // VOID PIERCE - Sword
+    VOID_PIERCE_RARE: { name: "§9Void Pierce", type: "sword", rarity: "Rare", min: 3, max: 5, description: "§7Shoots a void projectile that \n§7pierces through enemies, \n§7dealing §x{x} §7damage", cooldown: 80, scoreboard: "voidpierce" },
+    VOID_PIERCE_EPIC: { name: "§5Void Pierce", type: "sword", rarity: "Epic", min: 6, max: 8, description: "§7Shoots a void projectile that \n§7pierces through enemies, \n§7dealing §x{x} §7damage", cooldown: 80, scoreboard: "voidpierce" },
+    VOID_PIERCE_LEGENDARY: { name: "§6Void Pierce", type: "sword", rarity: "Legendary", min: 9, max: 14, description: "§7Shoots a void projectile that \n§7pierces through enemies, \n§7dealing §x{x} §7damage", cooldown: 80, scoreboard: "voidpierce" },
+    VOID_PIERCE_MYTHIC: { name: "§cVoid Pierce", type: "sword", rarity: "Mythic", min: 15, max: 25, description: "§7Shoots a void projectile that \n§7pierces through enemies, \n§7dealing §x{x} §7damage", cooldown: 80, scoreboard: "voidpierce" }
+};
